pyproject.toml
[project]
name = "tg-auto-test"
version = "0.1.0"
description = "Serverless Telegram bot testing library"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.6,<23",
    "telethon>=1.42.0,<2",
    "Pillow>=11.0,<13",
    "mutagen>=1.47.0,<2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10",
    "pytest-asyncio>=1.3.0,<2",
    "pytest-xdist>=3.8.0,<4",
    "ruff>=0.15.0,<0.16",
    "pylint>=4.0.4,<5",
    "vulture>=2.14.0,<3",
    "build>=1.2.0,<2",
    "twine>=6.0.0,<7",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["tg_auto_test*"]
exclude = ["tests*"]
